+++ src/dune_rules/ocaml_stdlib.ml
@@ -3,7 +3,7 @@ open! Import
 type t =
   { modules_before_stdlib : Module_name.Set.t
   ; exit_module : Module_name.t option
-  ; internal_modules : Glob.t
+  ; internal_modules : Predicate_lang.Glob.t
   ; loc : Loc.t
   }
 
@@ -24,7 +24,8 @@ let decode =
        field "modules_before_stdlib" (repeat Module_name.decode) ~default:[]
      and+ exit_module = field_o "exit_module" Module_name.decode
      and+ internal_modules =
-       field "internal_modules" Glob.decode ~default:Dune_lang.Glob.empty
+       field "internal_modules" Predicate_lang.Glob.decode
+         ~default:Predicate_lang.false_
      and+ loc = loc in
      { modules_before_stdlib = Module_name.Set.of_list modules_before_stdlib
      ; exit_module